@@ -1,10 +1,10 @@
 name: characters
-version: 1.0.0
+version: 1.1.0-dev
 description: String replacement with operations that are Unicode/grapheme cluster aware.
 homepage: https://www.github.com/dart-lang/characters
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.9.0-4.0.dev <2.10.0'
 dev_dependencies:
   test: "^1.6.0"
   pedantic: ^1.9.0
